+++ package.json
@@ -1,0 +1,31 @@
+{
+  "name": "awesome-book-es6",
+  "version": "1.0.0",
+  "description": "<a name=\"readme-top\"></a>\r <div align=\"center\">\r   <h3><b>Awesome Books</b></h3>\r </div>",
+  "main": "script.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/bhushan354/Awesome-Book-Es6.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/bhushan354/Awesome-Book-Es6/issues"
+  },
+  "homepage": "https://github.com/bhushan354/Awesome-Book-Es6#readme",
+  "devDependencies": {
+    "babel-eslint": "^10.1.0",
+    "eslint": "^7.32.0",
+    "eslint-config-airbnb-base": "^14.2.1",
+    "eslint-plugin-import": "^2.28.0",
+    "hint": "^7.1.10",
+    "stylelint": "^13.13.1",
+    "stylelint-config-standard": "^21.0.0",
+    "stylelint-csstree-validator": "^1.9.0",
+    "stylelint-scss": "^3.21.0"
+  }
+}
