--- style.css
@@ -1,196 +1,0 @@
-* {
-    margin: 0;
-    padding: 0;
-    font-family: Georgia, 'Times New Roman', Times, serif;
-  }
-  
-  .main {
-    background-image: url(./wallpaper.jpeg);
-    background-repeat: no-repeat;
-    background-size: cover;
-    box-shadow: inset 0 0 0 2000px rgba(189, 175, 197, 0.3);
-  }
-  
-  .navigation {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 16px;
-    margin: 16px;
-    border: 3px solid violet;
-  }
-  
-  .navigation ul {
-    display: flex;
-    list-style-type: none;
-  }
-  
-  .navigation ul li {
-    border-right: 2px solid violet;
-    padding: 0 8px;
-  }
-  
-  .navigation ul li a {
-    text-decoration: none;
-    color: blueviolet;
-  }
-  
-  .navigation ul li a:hover {
-    color: rgb(237, 134, 255);
-  }
-  
-  .navigation h3 {
-    color: blueviolet;
-    text-align: center;
-  }
-  
-  #library {
-    height: 83.5vh;
-  }
-  
-  section h1,
-  h2 {
-    color: white;
-    padding: 15px;
-    text-align: center;
-    padding-top: 14%;
-    margin-top: -33px;
-  }
-  
-  #book-list {
-    display: grid;
-    margin: 16px;
-    border: 3px solid violet;
-    position: absolute;
-    left: 13%;
-    width: 70%;
-  }
-  
-  .lists {
-    width: 100%;
-    display: flex;
-    justify-content: space-between;
-    padding: 5px 0;
-  }
-  
-  .lists p {
-    font-size: 16px;
-    padding: 7px;
-  }
-  
-  .lists span {
-    font-style: italic;
-    font-size: 18px;
-    font-weight: 600;
-  }
-  
-  .submit,
-  #btn {
-    font-size: 15px;
-    box-shadow: 2px 2px 3px grey;
-    background-color: rgb(255, 233, 255);
-    padding: 5px 10px;
-    border-radius: 6px;
-  }
-  
-  .submit {
-    margin: 2px 8px;
-  }
-  
-  .submit:hover,
-  #btn:hover {
-    background-color: rgb(237, 134, 255);
-  }
-  
-  .form {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    width: 90%;
-  }
-  
-  .contact-flex {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    margin-top: -33px;
-  }
-  
-  .book-form {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 17px;
-    margin-top: 58px;
-    padding-bottom: 17%;
-  }
-  
-  input#title,
-  input#author {
-    font-size: 18px;
-    padding: 7px 12px;
-    border: 1px solid violet;
-  }
-
-  
-.success {
-    color: white;
-    padding: 5px;
-    margin: 5px 0;
-    background-color: green;
-    position: absolute;
-    width: 14%;
-    left: 42.5%;
-    text-align: center;
-  }
-  
-  .error {
-    color: white;
-    padding: 5px;
-    margin: 5px 0;
-    background-color: red;
-    position: absolute;
-    width: 14%;
-    left: 42.5%;
-    text-align: center;
-  }
-  
-  .lists:nth-child(odd) {
-    background-color: rgb(255, 196, 255);
-  }
-  
-  .contacttext {
-    background-color: white;
-    padding: 24px;
-    margin-top: 34px;
-    padding-left: 50px;
-    border: 3px solid violet;
-    width: 32%;
-    margin-bottom: 19%;
-  }
-  
-  .contact-list {
-    list-style: none;
-    padding-left: 10px;
-  }
-  
-  footer {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    padding: 16px;
-    border: 3px solid violet;
-    background-color: rgb(240, 128, 254);
-    color: white;
-    font-weight: 600;
-  }
-  
-  .time {
-    display: flex;
-    justify-content: flex-end;
-    margin: 0 16px;
-    font-weight: 500;
-    color: white;
-  }
